pyproject.toml
[project]
name = "fastapi-oauth"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "pyjwt>=2.10.1",
]
